+++ src/components/header/Alarm/index.tsx
@@ -2,7 +2,7 @@ import { RefObject } from 'react';
 
 import classNames from 'classnames/bind';
 
-import { SVGS } from '@/constants/images';
+import { SVGS } from '@/constants/index';
 
 import styles from './Alarm.module.scss';
