+++ src/components/header/HeaderProfile/index.tsx
@@ -2,7 +2,7 @@ import { RefObject } from 'react';
 
 import classNames from 'classnames/bind';
 
-import { SVGS } from '@/constants/images';
+import { SVGS } from '@/constants/index';
 
 import Avatar from '@/components/Avatar';
